rm Walkers, put walk in Utilities
