--- docs/src/lib/internals/walkers.md
@@ -1,5 +1,0 @@
-# Walkers
-
-```@autodocs
-Modules = [Documenter.Walkers]
-```
