+++ src/Documenter.jl
@@ -46,7 +46,6 @@ include("Anchors.jl")
 include("Documents.jl")
 include("Builder.jl")
 include("Expanders.jl")
-include("Walkers.jl")
 include("CrossReferences.jl")
 include("DocChecks.jl")
 include("Writers/Writers.jl")
