+++ src/Utilities/Utilities.jl
@@ -640,4 +640,51 @@ include("DOM.jl")
 include("MDFlatten.jl")
 include("TextDiff.jl")
 
+##############################################
+# walk (previously in the Walkers submodule) #
+##############################################
+"""
+$(SIGNATURES)
+
+Calls `f` on `element` and any of its child elements. `meta` is a `Dict` containing metadata
+such as current module.
+"""
+walk(f, meta, element) = (f(element); nothing)
+
+# Change to the docstring's defining module if it has one. Change back afterwards.
+function walk(f, meta, block::Markdown.MD)
+    tmp = get(meta, :CurrentModule, nothing)
+    mod = get(block.meta, :module, nothing)
+    mod ≡ nothing || (meta[:CurrentModule] = mod)
+    f(block) && walk(f, meta, block.content)
+    tmp ≡ nothing ? delete!(meta, :CurrentModule) : (meta[:CurrentModule] = tmp)
+    nothing
+end
+
+function walk(f, meta, block::Vector)
+    for each in block
+        walk(f, meta, each)
+    end
+end
+
+const MDContentElements = Union{
+    Markdown.BlockQuote,
+    Markdown.Paragraph,
+    Markdown.MD,
+}
+walk(f, meta, block::MDContentElements) = f(block) ? walk(f, meta, block.content) : nothing
+
+const MDTextElements = Union{
+    Markdown.Bold,
+    Markdown.Header,
+    Markdown.Italic,
+}
+walk(f, meta, block::MDTextElements)      = f(block) ? walk(f, meta, block.text)    : nothing
+walk(f, meta, block::Markdown.Footnote)   = f(block) ? walk(f, meta, block.text)    : nothing
+walk(f, meta, block::Markdown.Admonition) = f(block) ? walk(f, meta, block.content) : nothing
+walk(f, meta, block::Markdown.Image)      = f(block) ? walk(f, meta, block.alt)     : nothing
+walk(f, meta, block::Markdown.Table)      = f(block) ? walk(f, meta, block.rows)    : nothing
+walk(f, meta, block::Markdown.List)       = f(block) ? walk(f, meta, block.items)   : nothing
+walk(f, meta, block::Markdown.Link)       = f(block) ? walk(f, meta, block.text)    : nothing
+
 end
